lesson6 - CalculatorImpl with proxy (fixed)

+++ homeworks/src/main/java/lesson6/Calculator.java
@@ -1,5 +1,7 @@
 package lesson6;
 
+import java.math.BigInteger;
+
 public interface Calculator {
-    Integer calc(Integer x);
+    BigInteger calc(Integer x);
 }
